Fix view initialization for Kanban and calendar
