+++ resources/views/projects/calendar.blade.php
@@ -14,37 +14,9 @@
         <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
             <div class="bg-white dark:bg-gray-800 overflow-hidden shadow-sm sm:rounded-lg">
                 <div class="p-6 text-gray-900 dark:text-gray-100">
-                    <div id="calendar"></div>
+                    <div id="calendar" data-events-url="{{ route('projects.calendar.data', $project) }}"></div>
                 </div>
             </div>
         </div>
     </div>
-
-    @push('scripts')
-    <script type="module">
-        import { Calendar } from '@fullcalendar/core';
-        import dayGridPlugin from '@fullcalendar/daygrid';
-        import frLocale from '@fullcalendar/core/locales/fr';
-
-        const calendarEl = document.getElementById('calendar');
-        const calendar = new Calendar(calendarEl, {
-            plugins: [dayGridPlugin],
-            initialView: 'dayGridMonth',
-            locale: frLocale,
-            headerToolbar: {
-                left: 'prev,next today',
-                center: 'title',
-                right: 'dayGridMonth,dayGridWeek'
-            },
-            events: '{{ route('projects.calendar.data', $project) }}',
-            eventClick: function(info) {
-                info.jsEvent.preventDefault();
-                if (info.event.url) {
-                    window.open(info.event.url, "_self");
-                }
-            }
-        });
-        calendar.render();
-    </script>
-    @endpush
 </x-app-layout>
